change in some models for addtweet services

+++ src/main/java/com/greenblog/repository/TweetRepo.java
@@ -11,5 +11,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TweetRepo extends CrudRepository<Tweet, Integer> {
+public interface TweetRepo extends CrudRepository<Tweet, Long> {
 }
